+++ scripts/__init__.py
@@ -2,3 +2,4 @@ from .generate_key import generate_key
 from .anonymize_gradebook import gradebook_anonymize
 from .anonymize_survey import survey_anonymize
 from .anonymize_ta_survey import anonymize_ta_survey
+from .clustering import prepare_data_for_clustering, clean_background_survey, cluster_survey
