+++ scripts/clustering/background_survey_cleaning/__init__.py
@@ -1,0 +1,1 @@
+from .background_survey_cleaning import clean_background_survey
