+++ util/__init__.py
@@ -1,1 +1,2 @@
 from .encoder import Encoder, EncoderException
+from .cluster import inertia_graph, kmeans_clustering, get_groups
